Stop using EventIdHash to compare messages in telemetry logs (#6526)
## Summary of changes

Compare the full log message and stack trace when determining uniqueness

## Reason for change

We were using a computed hash value to compare two log messages for
uniqueness. There's no good reason for that really, and we were getting
some hash collisions which was causing flakiness in tests (and
potentially, in real life). This should resolve the flakiness

## Implementation details

Use a `readonly record struct` to hold both the message and the stack
trace as the dictionary key.

## Test coverage

Covered by existing tests

## Other details

Would like to have used `ValueTuple` but we can't have nice things with
.NET FX 4.6.1