Build the test sample projects in a separate CI stage (#6498)
## Summary of changes

Simplify the sample build, and move the building of the samples to a
separate stage.

## Reason for change

Currently, the samples are built at the start of each integration stage
(i.e. for each OS/arch/framework). This typically takes ~10+ minutes,
depending on the details. If a test fails due to flake, the whole job
must be restarted, which means you must pay that cost again.

Additionally, building the samples at this point puts them in the
critical path for the whole CI run. By moving them to a separate stage,
which can start _immediately_, we can have the samples built ready by
the time the integration test stage starts

Finally, this is an incremental step in attempting to overhaul the Nuke
build by reducing the number of different targets, and the complexity of
the build in general.

## Implementation details

One of the key factors is creating a dedicated (autogenerated) "samples"
solution, that you can just run `dotnet build` on. This _should_ be a
lot faster than iterating through all the projects to find them, but it
also just simplifies the build stage in general. It's built rom the
"main" Datadog.Trace.sln, and can easily be updated by running
`.\tracer\build.ps1 RegenerateSolutions`. I have some other ideas around
this approach

> Note that I looked at trying to use
[Microsoft.Build.Traversal](https://github.com/microsoft/MSBuildSdks/blob/main/src/Traversal/README.md)
but I just couldn't get it to work on our solution, and this seemed
easier.

In the pipeline, I chose to build _all_ the samples for _all_ TFMs (in
parallel). This mostly for simplicity, as some stages test other TFMs to
the "default ones" (e.g. arm64 tests .NET 5 on PRs), but we can look to
improve that later. The sample build runs on Windows, as I suspected
that to give the fewest issues longer term. I also added a dedicate
macos sample build to make sure we don't commit a sample that _only_
builds on Windows! 😄

The biggest problem was CI Visibility. There some weirdness that happens
with the _Microsoft.TestPlatform.TestHost_ package. Namely, that on
Windows, the package copies an arch-specific binary into the output
folder. Unfortunately, this causes the tests to fail when it runs on
arm64 (i.e. a different arch). I spent a long time trying to work around
this, and in the end, just disabled the default copying for Windows. It
looks like the CI Vis tests don't currently run on Windows (⚠️
@tonyredondo) so this "works" but will need to be addressed somehow
later if we want to run them on Windows (we should).

Finally, I removed a lot of the old Nuke targets around building
samples. This isn't complete yet, but is a step in the right direction I
think.

## Test coverage

I've done a lot of runs of the CI, including checking it still works for
ASM and debugger etc. [This is one of the latest
runs](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=170878&view=results).

I have also tried to keep the local dev experience the same for now,
i.e. `BuildAndRunWindowsIntegrationTests` etc still work the same as
they did before.

## Other details

There's a lot of additional work and caveats about this work:

- The fully built samples are _big_: 20GB. We probably could/should dig
into that at some point...
- Currently the default samples (i.e. not the multi-version samples) are
only _built_ not _published_. That means they _don't_ contain all their
dependencies, and rely on the packages folder.
  - We _don't_ currently restore that folder between stages.
- That doesn't matter currently, because we're doing a "full solution"
restore in the build stage, and _are_ restoring it there
- In the (near) future I'd like to stop doing the global restore (and
simplify the "normal projects" build using a similar generated sln). At
that point we'll need to look into handling the samples differently
(either restoring the packages folder, or doing a publish instead of
build)
- The samples _don't_ depend on `Datadog.Trace` which is why we can take
this approach. However, some _do_ depend on `Datadog.Trace.Manual`.
Consequently, that project is rebuilt in the samples stage. It's not
ideal, but not a big deal IMO
- We don't build _all_ the samples that we currently reference in the
project. Some of this is for simplicity (e.g. aspnet samples) and some
is because we simply don't test some of the samples (`ExpenseItDemo`).
- Some stages are still doing their "own" building of samples. This is
fine for now, but it would be nice to remove that rebuilding as we
further simplify the Nuke build in the future.
- We can now relatively easy start doing the multi-version testing on
Windows if we want, as the samples are already built and available.
- We need to start testing ci Visibility on Windows 😬 
- In general, you'll need to run `RegenerateSolutions` when you add a
new sample. We could/should add a github action that flags this for you,
but can do that in a follow up PR

https://datadoghq.atlassian.net/browse/AIT-6624
https://datadoghq.atlassian.net/browse/LANGPLAT-5