[ASM] Fix WebApi blocking on Response (#6488)
## Summary of changes

Fixed the blocking on Response on WebApi.

When inside the HttpTracingModule, in the end method, if we want to
block the request, the previous code would throw a BlockException, and
is catch bu the tracing module and not the user code.

## Tests

New tests have been added on WebApi to check if the blocking iss
triggered.

## Reason for change

[Jira
ticket.](https://datadoghq.atlassian.net/jira/software/c/projects/APPSEC/boards/2876?selectedIssue=APPSEC-56204)