Disable the locked tracer in the runner integration tests (#6532)
## Summary of changes

Disable the locked tracer globally in the runner integration tests.

## Reason for change

The locked tracer was already disabled in CiRunCommandTests, but it
turns out there are other tests that require to disable it (we just got
lucky with the ordering). Rather than discovering them one by one, let's
just disable it globally.

## Implementation details

Also added the proper xunit.runner.json file to have more info on the
running tests in the CI.

## Other details

Discovered while trying to randomize the order of the tests.