Parallelise appsec system tests in CI (#6496)
## Summary of changes

- Parallelise all `appsec` system-tests scenarios in the CI.
`appsec` group scenarios have been broke down to scenarios, and each
individual scenario are run in a specific job.

- Remove old scenario `LIBRARY_CONF_CUSTOM_HEADERS_SHORT` and
`LIBRARY_CONF_CUSTOM_HEADERS_LONG` from the running list. They are both
now linked to the `LIBRARY_CONF_CUSTOM_HEADER_TAGS` scenario (which was
added to the `additionalScenarios` list)

- Speed up run of system tests by building the weblogs and agent docker
images before the execution of the tests matrix.

The majority of all the scenarios get completed in less than 5 min.
From 1h before the parallelisation:

![image](https://github.com/user-attachments/assets/b26f55b3-eb5a-49a6-b359-ece749626a1c)

To approximately more than 45 min after (depending of the attribution of
runners). That's also the parametric tests that are taking now the more
time:

![image](https://github.com/user-attachments/assets/05bda4f7-4f7a-49f6-8bb1-7643431e96c6)

# Implementation details

### List all scenarios that are part of groups

This is using the following command to get the scenarios from a scenario
group name:
`PYTHONPATH=. python utils/scripts/compute-workflow-parameters.py dotnet
-g appsec`

This way we can dynamically get the available weblogs in the
system-tests and all scenarios.

### Build the matrix

In order to create the matrix, a python script have been created to
merge all scenarios for listed groups and the additional scenario that
we want to run that are not part of a group.

We grep the output of the system-test command to fully get all scenarios
and construct back a json dictionary with all the scenario name and the
weblog variant that need to be used with it.

### Update

For future update, if a new group scenario needs to be broke down into
scenarios, it should be added to this variable list:
`scenarioGroups: "appsec"`

If specific scenario or group scenario needs to run directly, they need
to be added to that list:
`additionalScenarios: '[\"PARAMETRIC\", \"REMOTE_CONFIG_SCENARIOS\",
\"TELEMETRY_SCENARIOS\", \"DEFAULT\", \"INTEGRATIONS\",
\"INTEGRATIONS_AWS\", \"CROSSED_TRACING_LIBRARIES\",
\"DEBUGGER_SCENARIOS\", \"+S PROFILING +S TRACE_PROPAGATION_STYLE_W3C +S
LIBRARY_CONF_CUSTOM_HEADER_TAGS\"]'
`
### Build time

The weblog and agent docker images are build before and given as
artefacts for all test jobs to reduce the completion time.

## Reason for change

The system-tests were timing out in a bunch of PRs. Some group scenarios
took more than 1h to be run.