Fix pinned locals parsing + bump dotnet version for protobuf and eShopOnWeb exploration tests (#6475)
## Summary of changes

- fixed IL instrumentation error related to pinned variables. When we
parsed the LocalVarSig into an array of locals, we didn't handle pinned
variables. It caused the variables following it to be parsed
incorrectly, leading to type confusion (e.g treating `Span<T>` as
`system.Threading.Thread`) which caused bad instrumentation decision to
be made. One of them is the usage of by-ref-like types as generic param
(to be used by the debugger invokers) and failed with
`InvalidProgramException` because by-ref-like is not allowed to be used
as generic param.
- change the version of the dotnet framework used to run 2 exploration
tests from netcore 2.1 to a more recent one
also change the version of those libraries used to have one that's
compatible with more recent dotnet versions (this is what revealed the
pb mentioned above)

## Reason for change

those tests were using netcore 2.1 which we don't actively support
anymore, so it makes sense to test them with a supported version.

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>
Co-authored-by: Matan Green <matangrn@gmail.com>
Co-authored-by: dudik <dudi.keleti@gmail.com>