Move SpanContextPropagator to TracerManager (#6511)
## Summary of changes

Store the `SpanContextPropagator` on the `TracerManager` instead of
using it as a singleton.

## Reason for change

`SpanContextPropagator` was used as a singleton (through
`SpanContextPropagator.Instance`) even though it is dependent on the
current tracer configuration. Because of that, some unit tests were
impacting other tests despite using a scoped tracer.

## Implementation details

Added a `SpanContextPropagator` property on `TracerManager` and updated
all usages. When possible, I reused the available instance of tracer,
otherwise I used `TracerManager.Instance`.

In the process I discovered that `DistributedContextExtractor` isn't
used anymore so I removed it.

## Other details

Discovered while trying to randomize the order of the tests.