MVP for test containers (with aerospike) (#5031)
## Summary of changes

Add testcontainers to our integration tests. Convert the Aerospike tests
to use testcontainers.

## Reason for change

Much easier to launch the tests locally.

## Implementation details

For each service to run in docker, an instance of `ContainerFixture`
must be implemented to explain what image to use and how to configure
it. Then tests that depend on it must import the fixture.

In the xunit executor, we find all the docker fixtures needed to run the
set of tests, and we start them. They're cleaned up after all tests
cases have run (for future improvement, we can try to stop the docker
images as soon as they're not needed anymore).
Because fixtures are not singleton, the ContainersRegistry is used to
store the objects across multiple fixtures.

## Test coverage

Only Aerospike is converted in this PR.

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>