Randomize the order of tests (#5831)
## Summary of changes

Randomize the order of the tests.

## Reason for change

Flaky tests are much harder to fix when we discover them long after they
have been written. By randomizing the order of the tests, I'm hoping to
make them fail earlier.
In practice, this could temporarily increase the overall flakiness, but
I expect this will reduce the overall effort spent on fixing tests.

## Implementation details

In `CustomTestFramework`, randomize the list of all tests in each
collections, and the collections themselves.
The seed is displayed in the output. When a test order causes tests to
fail, this allows to deterministically reproduce that test order.

## Other details

Four other issues were found thanks to this:
https://github.com/DataDog/dd-trace-dotnet/pull/6535,
https://github.com/DataDog/dd-trace-dotnet/pull/6532,
https://github.com/DataDog/dd-trace-dotnet/pull/6511,
https://github.com/DataDog/dd-trace-dotnet/pull/6509