Disable profiling in smoke tests (#6548)
## Summary of changes

Disables profiling in the crash smoke tests

## Reason for change

There's a freeze we sometimes hit in the smoke tests. It was similar to
an issue we saw in some integration tests. Disabling the profiler in
those tests (in https://github.com/DataDog/dd-trace-dotnet/pull/6497)
resolved the issue, so we think it will resolve it here. We'll
investigate the freeze separately, this is just trying to reduce flake.

## Implementation details

Explicitly set `DD_PROFILING_ENABLED=0`

## Test coverage

This is the test