Don't run integration tests on `netcoreapp2.1` (#6560)
## Summary of changes

Disables the `netcoreapp2.1` integration tests

## Reason for change

#6498 broke the .NET Core 2.1 integration tests, due to a strange
"missing dependency" issue. I believe this is due to the somewhat
frankenstein approach we have to building and restoring right now. I
hope to improve this in the near future, but fixing it for 2.1 right now
is too big of a cost.

## Implementation details

Remove 2.1 from the tests

## Test coverage

A lot less now unfortunately, but it's not "supported" anyway

## Other details

This is the quickest, easiest, way to remove the 2.1 tests to fix the
build. Later we can decide if this is the permanent solution we want (in
which case we can stop _building_ for 2.1 too) or if we want/can
reenable some of the basic tests at least.