Don't log a warning for CORPROF_E_PROFILER_CANCEL_ACTIVATION (#6550)
## Summary of changes

Don't log a warning if the classfactory of the tracer or the profiler
returns CORPROF_E_PROFILER_CANCEL_ACTIVATION.

## Reason for change

The tracer and the profiler return this error code when explicitly
disabled (either by an `DD_{0}_ENABLED` setting, or by process
exclusion). This is expected and therefore should be logged as info
rather than warning.

## Implementation details

Just check the HResult. As a bonus, removed a superfluous log.

## Test coverage

Tested on my machine.