Add support for RabbitMQ v7 (#6479)
## Summary of changes

Adds support for RabbitMQ v7

## Reason for change

We want to support the latest versions of packages. 7.x.x was released
~6 weeks ago, with _significant_ changes to both the public API and the
internal API (which is probably why it hasn't been explicitly requested
yet).

## Implementation details

This one took a _lot_ of work. In summary:

- Basically all the integrations had to be rewritten to work with the
new async APIs
- Where possible, simply called into the original sync integration
(other than where the work is trivial)
- Some challenges, in that the duck types had to change.
- Most problematic was the `BasicProperties` because now
`IReadOnlyBasicProperties` is null, which means we can't easily add the
headers if they're not provided (to propagate context).
- Worked around it with an additional instrumentation, but I'm not super
happy about it 😅
- Had to completely rewrite the sample app to use the new async APIs
- Originally I tried to `#if RABBITMQ_7_0` my way in the existing
methods, but was impossible to follow, so split into -pre and -post v7
methods instead.
- Tweaked some of the pre-V7 methods in the sample to be async to
improve reuse

## Test coverage

We get the same overall snapshots for v7 as we do for v6, so 🤞 Did an
"all TFMs, all versions" run
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=170428&view=results)
(and a previous one for DSM)

## Other details

Stacked on:
- https://github.com/DataDog/dd-trace-dotnet/pull/6463
- https://github.com/DataDog/dd-trace-dotnet/pull/6480

> [!NOTE]
> This PR actually highlighted a current limitation in our CallTarget
instrumentation. If an API uses `ValueTask`, which comes from a
_library_ rather than being built-in (i.e. it's netfx or < .NET Core
3.1) then we _don't_ instrument it correctly. Originally I worked around
it, but there's a "proper" fix in #6480. Assuming that actually works
and is merged, then we don't need the workaround, otherwise I can
re-enable it.