[Profiler] Fix Windows 64 bit stack walking (#6583)
## Summary of changes
Fix Windows 64 bit stack walking code

## Reason for change
On Windows, 64 bit stack walking failed (maybe after a Visual Studio
update). After investigating, it was due to fact that the execution
context was retrieved in a function that was inlined before. Since this
context is then used in the same method to iterate on different frames,
then the stack should not be updated (i.e. not returning from a function
call).
The existing code should have always failed but a previous version of
the compiler probably decided to inline it. And this is no more the case
for VS 17.12.

## Implementation details
Move the context retrieval code in the same function that walk the stack

## Test coverage
Already covered in all providers that walk the stack

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->