LLM PR reviewer (#6381)
## Summary of changes

This PR adds one stage to the pipeline that writes a report with the
code changes of the PR by sending them to OpenAI for a code review. This
stage is optional and is only launched when the pipeline variable
"generate_llm_report" is set to "true"
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build?definitionId=54).
By default, this variable will be set as "false" unless it's decided
that we want to add these kind of reports by default.

Also, a report can be generated locally by running the task LLMPRReview
For example: 
`tracer\build LLMPRReview -GITHUB_TOKEN <GH_TOKEN> -OPEN_AI_KEY
<OPEN_AI_KEY> -PullRequestNumber <PRNumber>`

This task, when run in local mode, will generate two files: 
* Changes.txt: a file containing the prompt sent to OpenAI
* Results.txt: the LLM report provided by the OpenAI API call

## Reason for change

It's an innovation week project.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>