Refactor the NGEN tracking fix to be more performant (#6598)
## Summary of changes

This PR refactor the NGEN tracking fix to remove the `lock` and the
`unordered_set` and reuse the `module_ids` lock and a vector + 3
mdMethodDef field.

## Reason for change

After #6588 was merged, @andrewlock saw some performance impact in the
Execution Benchmarks.

## Implementation details

Remove a lock and an unordered_set for just a field comparison + a
vector Contains if a field is found.

This should improve everything on the missed hit because is just a
comparison over 3 fields.

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->