[Exception Replay] Fixed capturing issue of async methods with `await` in finally block + added missing snapshot attributes + better frame matching algorithm (#6549)
## Summary of changes
- There was an issue where certain exceptions were not captured due to
misleading stack traces. The API
`ExceptionDispatchInfo.Capture(ex).Throw()` is used (mostly by the .NET
compiler although it's a public API that can also be used directly) in
async methods to capture exceptions that may be rethrown on another
thread. The compiler emits this call when there's `await` in a `finally`
block. There are certain code constructs that result with `await` in
`finally` such as `await using` and `await foreach` statements.
- In those situations it was difficult to capture an exception stack
trace because the stack trace being analyzed for instrumentation was
incorrect. It showed a causality chain that has not really happened
(duplicated frames that are not really resulted in recursion call but
because the exception was freezed & unfreezed for rethrow). To be able
to overcome that, the algorithm was improved to realize if a method is
'misleading' (duplicated in the stack trace due to
`ExceptionDispatchInfo` usage by analyzing method's bytecode) or not and
handling the enter/leave capturing accordingly.
- The matching algorithm between the exception stack trace and the
snapshots was also improved. It's now based on the string representation
of the exception instead of solely `new StackTrace(ex)`. It mitigates
mismatches between a frame and a snapshot for better user experience.
- Added the following missing attributes on snapshots produced from
Exception Replay:
- `exceptionHash`: represents the hash of the exception, should be
identical among snapshots produced for similar exceptions in the same or
different occurrence.
- `exceptionId`: a UUID that is being generated when an exception is
captured. It's identical among the snapshots that were uploaded together
for an exception for the same occurrence.
- `frameIndex`: the index, in `exception.ToString()`, where current
snapshot is part of (indexed from `0` to the outer most user frame
index, where `0` represents the inner most frame, e.g the throwing
method).
 
## Reason for change
- Improve capturing accuracy for exception stack traces that holds
misleading duplicated frames.
- Matching frame <-> snapshot more accurately, based on the actual
`exception.ToString()` instead of `new StackTrace(ex)`.

## Implementation details
- Analyzing methods bytecode to determine if a method calls
`ExceptionDispatchInfo.Capture`.
- Improved the instrumentation logic to, at runtime, accommodate for
potential false duplicated frames.
- The indices of frames as span tags represent the actual location in
the `exception.ToString` to avoid frame <-> snapshot confusing
mismatches.

## Test coverage
- New Exception Replay tests were added that previously couldn't be
captured.
- `ILAnalyzerTests` that validate the functionality of the `ILAnalyzer`
class that was added to determine if a given method's IL has a call to
`ExceptionDispatchInfo.Throw`.
- `MethodMatcherTests` a bunch of tests that make sure the new class
`MethodMatcher` works properly. This class was added to check if a
string representation of a method (taken from `exception.ToString`)
matches a given `MethodBase` instance.

## Fixes
DEBUG-2789