Update dotnet-install scripts for impending change of .NET install domains (#6484)
## Summary of changes

Updates the `dotnet-install.sh` and `dotnet-install.ps` scripts
following advice in https://github.com/dotnet/announcements/issues/336

## Reason for change

From the linked announcement:

> Some .NET binaries and installers are hosted on Azure Content Delivery
Network (CDN) domains that end in .azureedge.net. These domains are
hosted by [edg.io](https://edg.io/), which will [soon cease
operations](https://learn.microsoft.com/en-us/azure/cdn/edgio-retirement-faq)
due to bankruptcy. We are required to [migrate to a new
CDN](https://learn.microsoft.com/azure/frontdoor/migrate-cdn-to-front-door)
and will be using new domains going forward. It is possible that
.azureedge.net domains will have downtime or become permanently
unavailable.

## Implementation details

Update the scripts with the versions provided in the initial
announcements. These scripts aren't "final", and we don't need to merge
them _yet_ (and we will need to update the VMs when we do), so for now
this is a "break glass if things stop working" scenario. Not that the
powershell script isn't signed, so we should definitely wait for that
too if we can.

## Test coverage

This PR is the test - as long as it builds ok, we're good

## Other details

I haven't really looked through the changes. They're _mostly_
superficial, other than preferring the new CDN domain and using a
fallback