[ASM] [IAST] Enable email injection unit tests (#6595)
## Summary of changes

We disabled the email injection unit tests in the past because they were
causing some problems under netcore 3.1. The tests would hang when
trying to connect a non existing SMTP server. The tests were disabled to
avoid this situation. Now, the tests are enabled again and no smtp
server or email address are provided, which throws an exception without
going further but we are still able to detect the vulnerability. This
prevents the previous situation.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->