Move the Nuke arguments into the gitlab file (#6611)
## Summary of changes

Allow specifying the Nuke targets to run in the gitlab file

## Reason for change

Currently they're baked into the build image. That means if we need to
change the args, we need to build a whole new image

## Implementation details

Pass the arguments as commands in the `docker run` call. These get
passed as arguments to the `entrypoint.bat` script. The script then
reads the args and splats them into the Nuke call.

## Test coverage

I've tested the splatting, but an end-to-end test requires rebuilding
the image. The good news is that this change works with the old image
(it's a no-op) so it's safe to merge anyway.

## Other details

That docker line was getting a bit hefty, so split it for "readability"