[IAST] Fix flaky Interpolated Strings tests (#6555)
## Summary of changes
Fix test flakiness by keeping alive the tainted boxed object during the
lifespan of the ref struct

## Reason for change
If a GC collect happened between the tainting of the
`DefaultInterpolatedStringHandler` and the call to `ToString` the boxed
object was collected and the weak reference was lost, causing the loss
of the tainted status.
Thanks to @e-n-0 for finding what was happening

IDK who said flakiness was bad

## Implementation details
Added a thread local queue to hold the boxed object of the tainted ref
structs

## Test coverage
Added a unit test

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Flavien Darche <flavien.darche@datadoghq.com>