Stop intalling all the .NET SDK versions in benchmark agents (#6612)
## Summary of changes

Only install the SDK version specified in the global.json

## Reason for change

Currently the benchmark agents are installing _every_ patch release of
_every_ TFM we ever test against. That means they slowly run out of
space. We're only (currently) running benchmarks against EOL TFMs
anyway, so there _aren't_ new versions of these to test against anyway,
so this is purely just a maintenance burden.

## Implementation details

Use `install-latest-dotnet-sdk.yml` instead of `install-dotnet-sdks.yml`

## Test coverage

Meh, I'm sure it'll be fine. 

## Other details

Yeah "latest" is not a great description for the task but ah well 🙈