[Tracing] Add configuration to change the trace context propagation extract behavior (APMAPI-1009) (#6320)
This implements two features within the context extraction logic:
- Adds a new `DD_TRACE_PROPAGATION_BEHAVIOR_EXTRACT` to update the logic
of context extraction independently of the propagator list. The
behaviors modify the returned SpanContext and SpanLinks fields to better
handle incoming contexts that could cause issues for service owners
- When additional trace contexts are identified and their trace-id does
not match the first extracted trace context, create a span link that
represents that span context. This is a missing feature that has been
standardized and adopted across libraries and it is extremely adjacent
to the first feature, so it was done in this effort