[CI Visibility] Disable ImpactedTests flaky test (#6616)
## Summary of changes
This PR disables a test failing in Alpine due to a Detached Head status.
We are working in a way to handle this situation.

## Reason for change
Test failing in Alpine

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->