[ASM] Fix Blocking Redirect Invalid Status Code (#6619)
## Summary of changes

This PR fixes the returned status code of the response when a request
has been blocked. The configured status code from the blocking action
was not applied. It was always `302` or `301` that was applied on the
response.

## Reason for change

Incorrect status code value applied.
(Detected by System Tests: the test wasn't passing).

## Implementation details

The `Redirect()` method from either the .NET Framework or .NET Core sets
the status code when called. But in our implementation we were setting
the blocking action status code and after we called that method, thus
overriding the status code.

## Test coverage

Integrations tests have been implemented for both .NET Framework and
.NET Core.